Rename serializers
Change Log:
- Rename *_serializers -> *
- Allocate one line for each field.
- Remove unnecessary codes.
