+++ post/serializers/__init__.py
@@ -2,10 +2,10 @@ from rest_framework import serializers
 
 from user.models import User
 
-from .category_serializers import *  # noqa 401
-from .comment_serializers import *  # noqa 401
-from .post_serializers import *  # noqa 401
-from .tag_serializers import *  # noqa 401
+from .category import *  # noqa 401
+from .comment import *  # noqa 401
+from .post import *  # noqa 401
+from .tag import *  # noqa 401
 
 
 class TempUserSerializer(serializers.ModelSerializer):
