+++ post/serializers/comment.py
@@ -1,6 +1,3 @@
-from typing import Any
-
-# from django.db.models.manager import Manager
 from rest_framework import serializers
 
 from post.models import Comment
@@ -14,14 +11,16 @@ class CommentSerializer(serializers.ModelSerializer):
         required=False
     )
 
-    def get_queryset(self) -> Any:
-        return Comment.objects.filter(is_active=True)
-
     class Meta:
         model = Comment
         depth = 1
-        fields = ('id', 'content', 'post', 'createdDate',
-                  'parentComment', 'author', 'is_active')
+        fields = ('id',
+                  'content',
+                  'post',
+                  'createdDate',
+                  'parentComment',
+                  'author',
+                  'is_active')
 
 
 class CommentCreateSerializer(serializers.ModelSerializer):
@@ -33,7 +32,11 @@ class CommentCreateSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = Comment
-        fields = ('id', 'content', 'post', 'parentComment', 'author')
+        fields = ('id',
+                  'content',
+                  'post',
+                  'parentComment',
+                  'author')
 
 
 class CommentPutSerializer(serializers.ModelSerializer):
