+++ post/serializers/post.py
@@ -5,7 +5,7 @@ from rest_framework import serializers
 
 from post.models import Comment, Post
 
-from .comment_serializers import CommentSerializer
+from .comment import CommentSerializer
 
 T = TypeVar('T', bound=models.Model)
 
@@ -17,9 +17,17 @@ class PostSerializer(serializers.ModelSerializer):
     class Meta:
         model = Post
         depth = 1
-        fields = ('id', 'title', 'content', 'location', 'capacity',
-                  'date', 'timeOfDay', 'createdDate', 'author',
-                  'category', 'tags')
+        fields = ('id',
+                  'title',
+                  'content',
+                  'location',
+                  'capacity',
+                  'date',
+                  'timeOfDay',
+                  'createdDate',
+                  'author',
+                  'category',
+                  'tags')
 
 
 class PostDetailSerializer(serializers.ModelSerializer):
@@ -35,9 +43,18 @@ class PostDetailSerializer(serializers.ModelSerializer):
     class Meta:
         model = Post
         depth = 1
-        fields = ('id', 'title', 'content', 'location', 'capacity',
-                  'date', 'timeOfDay', 'createdDate', 'author',
-                  'category', 'tags', 'comments')
+        fields = ('id',
+                  'title',
+                  'content',
+                  'location',
+                  'capacity',
+                  'date',
+                  'timeOfDay',
+                  'createdDate',
+                  'author',
+                  'category',
+                  'tags',
+                  'comments')
 
 
 class PostCreateSerializer(serializers.ModelSerializer):
@@ -45,8 +62,16 @@ class PostCreateSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = Post
-        fields = ('id', 'title', 'content', 'location', 'capacity',
-                  'date', 'timeOfDay', 'author', 'category', 'tags')
+        fields = ('id',
+                  'title',
+                  'content',
+                  'location',
+                  'capacity',
+                  'date',
+                  'timeOfDay',
+                  'author',
+                  'category',
+                  'tags')
 
 
 class PostPatchSerializer(serializers.ModelSerializer):
@@ -60,5 +85,11 @@ class PostPatchSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = Post
-        fields = ('id', 'title', 'content', 'location', 'capacity',
-                  'date', 'timeOfDay', 'tags')
+        fields = ('id',
+                  'title',
+                  'content',
+                  'location',
+                  'capacity',
+                  'date',
+                  'timeOfDay',
+                  'tags')
